[INLONG-8239][Sort] Extract common code
