+++ inlong-sort/sort-formats/format-inlongmsg-base/src/main/java/org/apache/inlong/sort/formats/inlongmsg/InLongMsgMixedValidator.java
@@ -18,12 +18,10 @@
 package org.apache.inlong.sort.formats.inlongmsg;
 
 import org.apache.inlong.sort.formats.base.TableFormatConstants;
+import org.apache.inlong.sort.formats.util.ValidateDescriptorString;
 
 import org.apache.flink.table.descriptors.DescriptorProperties;
 import org.apache.flink.table.descriptors.DescriptorValidator;
-
-import static org.apache.inlong.sort.formats.base.TableFormatConstants.FORMAT_PROPERTY_VERSION;
-import static org.apache.inlong.sort.formats.base.TableFormatConstants.FORMAT_TYPE;
 
 /**
  * Validator for mixed inlongmsg formats.
@@ -32,8 +30,7 @@ public class InLongMsgMixedValidator implements DescriptorValidator {
 
     @Override
     public void validate(DescriptorProperties properties) {
-        properties.validateString(FORMAT_TYPE, false, 1);
-        properties.validateString(FORMAT_PROPERTY_VERSION, true, 1);
+        properties = ValidateDescriptorString.getValidateProperties(properties);
         properties.validateString(TableFormatConstants.FORMAT_DELIMITER, true, 1, 1);
         properties.validateString(TableFormatConstants.FORMAT_ENTRY_DELIMITER, true, 1, 1);
         properties.validateString(TableFormatConstants.FORMAT_KV_DELIMITER, true, 1, 1);
